+++ path/path.go
@@ -16,7 +16,7 @@ import (
 	"path/filepath"
 	"strings"
 
-	"pkg.re/essentialkaos/ek.v2/env"
+	"pkg.re/essentialkaos/ek.v3/env"
 )
 
 // ////////////////////////////////////////////////////////////////////////////////// //